--- NBA Comparison/ContentView.swift
@@ -1,23 +1,0 @@
-//
-//  ContentView.swift
-//  NBA Comparison
-//
-//  Created by chuck on 7/24/20.
-//  Copyright © 2020 AlexZ. All rights reserved.
-//
-
-import SwiftUI
-
-struct ContentView: View {
-    var body: some View {
-        NavigationView {
-            SelectPlayerView()
-        }
-    }
-}
-
-struct ContentView_Previews: PreviewProvider {
-    static var previews: some View {
-        ContentView()
-    }
-}
