+++ NBA Comparison/NBA Comparison/BallDontLie.swift
@@ -69,25 +69,25 @@ class BallDontLie {
     
     static let seasonAverageStats = [
         "games_played": Stats.GAMES_PLAYED,
-        "min": Stats.MIN,
-        "fgm": Stats.FGM,
-        "fga": Stats.FGA,
-        "fg3m": Stats.FG3M,
-        "fg3a": Stats.FG3A,
-        "ftm": Stats.FTM,
-        "fta": Stats.FTA,
-        "oreb": Stats.OREB,
-        "dreb": Stats.DREB,
-        "reb": Stats.REB,
-        "ast": Stats.AST,
-        "stl": Stats.STL,
-        "blk": Stats.BLK,
-        "turnover": Stats.TURNOVER,
-        "pf": Stats.PF,
-        "pts": Stats.PTS,
-        "fg_pct": Stats.FG_PCT,
-        "fg3_pct": Stats.FG3_PCT,
-        "ft_pct": Stats.FT_PCT,
+        "min": Stats.MINUTES,
+        "fgm": Stats.FIELD_GOALS_MADE,
+        "fga": Stats.FIELD_GOAL_ATTEMPTS,
+        "fg3m": Stats.THREE_POINTERS_MADE,
+        "fg3a": Stats.THREE_POINT_ATTEMPTS,
+        "ftm": Stats.FREE_THROWS_MADE,
+        "fta": Stats.FREE_THROW_ATTEMPTS,
+        "oreb": Stats.OFFENSIVE_REBOUNDS,
+        "dreb": Stats.DEFENSIVE_REBOUNDS,
+        "reb": Stats.REBOUNDS,
+        "ast": Stats.ASSISTS,
+        "stl": Stats.STEALS,
+        "blk": Stats.BLOCKS,
+        "turnover": Stats.TURNOVERS,
+        "pf": Stats.PERSONAL_FOULS,
+        "pts": Stats.POINTS,
+        "fg_pct": Stats.FIELD_GOAL_PERCENTAGE,
+        "fg3_pct": Stats.THREE_POINT_PERCENTAGE,
+        "ft_pct": Stats.FREE_THROW_PERCENTAGE,
     ]
     
     func loadSeasonAverages(season: Int, playerIds : [Int], completionHandler: @escaping ([Dictionary<Stats, StatValue>]?, Error?) -> Void) {
