+++ NBA Comparison/NBA Comparison/PlayerStore.swift
@@ -27,3 +27,9 @@ class PlayerStore : ObservableObject {
         }
     }
 }
+
+struct PlayerStore_Previews: PreviewProvider {
+    static var previews: some View {
+        /*@START_MENU_TOKEN@*/Text("Hello, World!")/*@END_MENU_TOKEN@*/
+    }
+}
